iota_indexer::restore

Function start

pub async fn start(
    network: Network,
    epoch: Option<u64>,
    staging_path: &Path,
    genesis_path: &Path,
    num_parallel_downloads: NonZeroUsize,
    pg_indexer_store: PgIndexerStore,
) -> IndexerResult<()>

Restores the indexer database from the formal snapshot for the given network and epoch.

This guarantees that the formal snapshot is verified, by comparing the root state hash of the live objects against the verified commitment of the network at the given epoch.

Errors

Returns an error if:

The reader cannot be instantiated.

The persist pipeline fails

The snapshot fails verification.