Module iota_indexer::schema::checkpoints::dsl

Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Structs

checkpoint_commitments

checkpoint_digest

checkpoints

The actual table struct

computation_cost

end_of_epoch

end_of_epoch_data

epoch

max_tx_sequence_number

min_tx_sequence_number

network_total_transactions

non_refundable_storage_fee

previous_checkpoint_digest

sequence_number

storage_cost

storage_rebate

timestamp_ms

total_gas_cost

tx_digests

validator_signature