iota_indexer::store

Module diesel_macro

Macros

blocking_call_is_ok_or_panic

Check that we are in a context conducive to making blocking calls. This is done by either:

read_only_blocking

Runs a blocking SQL query.

read_only_repeatable_blocking

run_query

Runs a blocking SQL query.

run_query_async

Runs an async SQL query wrapped in a spawn blocking task.

run_query_repeatable

Runs a blocking SQL query.

run_query_repeatable_async

spawn_read_only_blocking

Runs an async SQL query wrapped in a spawn blocking task.

transactional_blocking_with_retry

Runs a blocking SQL query.

Constants

CALLED_FROM_BLOCKING_POOL

Functions

mark_in_blocking_pool

Marks the current thread as being in a blocking pool.