iota_indexer::store::indexer_store

Trait IndexerStore

pub trait IndexerStore:
    Any
    + Clone
    + Sync
    + Send
    + 'static {

    // Required methods
    fn get_latest_checkpoint_sequence_number<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_available_epoch_range<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(u64, u64), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_available_checkpoint_range<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(u64, u64), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_latest_object_snapshot_watermark<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<CommitterWatermark>, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_latest_object_snapshot_checkpoint_sequence_number<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_chain_identifier<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_protocol_configs_and_feature_flags(
        &self,
        chain_id: Vec<u8>,
    ) -> Result<(), IndexerError>;
    fn persist_object_history<'life0, 'async_trait>(
        &'life0 self,
        object_changes: Vec<TransactionObjectChangesToCommit>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_object_versions<'life0, 'async_trait>(
        &'life0 self,
        object_versions: Vec<StoredObjectVersion>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_objects_snapshot<'life0, 'async_trait>(
        &'life0 self,
        object_changes: Vec<TransactionObjectChangesToCommit>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_checkpoints<'life0, 'async_trait>(
        &'life0 self,
        checkpoints: Vec<IndexedCheckpoint>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_transactions<'life0, 'async_trait>(
        &'life0 self,
        transactions: Vec<IndexedTransaction>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_optimistic_transaction_in_existing_transaction(
        &self,
        conn: &mut PgConnection,
        transaction: OptimisticTransaction,
    ) -> Result<(), IndexerError>;
    fn persist_events<'life0, 'async_trait>(
        &'life0 self,
        events: Vec<IndexedEvent>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_event_indices<'life0, 'async_trait>(
        &'life0 self,
        event_indices: Vec<EventIndex>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_displays<'life0, 'async_trait>(
        &'life0 self,
        display_updates: BTreeMap<String, StoredDisplay>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_packages<'life0, 'async_trait>(
        &'life0 self,
        packages: Vec<IndexedPackage>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_epoch<'life0, 'async_trait>(
        &'life0 self,
        epoch: EpochToCommit,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn advance_epoch<'life0, 'async_trait>(
        &'life0 self,
        epoch: EpochToCommit,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_network_total_transactions_by_end_of_epoch<'life0, 'async_trait>(
        &'life0 self,
        epoch: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn refresh_participation_metrics<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn as_any(&self) -> &dyn Any;
    fn persist_displays_in_existing_transaction(
        &self,
        conn: &mut PgConnection,
        display_updates: Vec<&StoredDisplay>,
    ) -> Result<(), IndexerError>;
    fn persist_objects_in_existing_transaction(
        &self,
        conn: &mut PgConnection,
        object_changes: Vec<TransactionObjectChangesToCommit>,
    ) -> Result<(), IndexerError>;
    fn update_watermarks_upper_bound<'life0, 'async_trait, E>(
        &'life0 self,
        watermark: CommitterWatermark,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where E::Iterator: Iterator<Item: AsRef<str>>,
             E: 'async_trait + IntoEnumIterator,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn update_watermarks_lower_bound<'life0, 'async_trait>(
        &'life0 self,
        watermarks: Vec<(PrunableTable, u64)>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_watermarks<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(Vec<StoredWatermark>, i64), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_checkpoint_objects<'life0, 'async_trait>(
        &'life0 self,
        objects: Vec<CheckpointObjectChanges>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_tx_global_order<'life0, 'async_trait>(
        &'life0 self,
        tx_order: Vec<TxGlobalOrder>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn persist_tx_indices<'life0, 'async_trait>(
        &'life0 self,
        indices: Vec<TxIndex>,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn prune_table_by_checkpoint_range<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table: &'life1 PrunableTable,
        min_checkpoint: u64,
        max_checkpoint: u64,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn prune_table_by_tx_range<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table: &'life1 PrunableTable,
        min_tx: u64,
        max_tx: u64,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn prune_table_by_global_seq_with_limit<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table: &'life1 PrunableTable,
        start: u64,
        end: u64,
        limit: i64,
    ) -> Pin<Box<dyn Future<Output = Result<usize, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_watermark_lowest_unpruned_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table: &'life1 PrunableTable,
        lowest_unpruned_key: u64,
    ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_watermark_by_entity<'life0, 'async_trait>(
        &'life0 self,
        entity: String,
    ) -> Pin<Box<dyn Future<Output = Result<Option<StoredWatermark>, IndexerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Required Methods

fn get_latest_checkpoint_sequence_number<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_available_epoch_range<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(u64, u64), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_available_checkpoint_range<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(u64, u64), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_latest_object_snapshot_watermark<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<CommitterWatermark>, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_latest_object_snapshot_checkpoint_sequence_number<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_chain_identifier<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_protocol_configs_and_feature_flags( &self, chain_id: Vec<u8>, ) -> Result<(), IndexerError>

fn persist_object_history<'life0, 'async_trait>( &'life0 self, object_changes: Vec<TransactionObjectChangesToCommit>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_object_versions<'life0, 'async_trait>( &'life0 self, object_versions: Vec<StoredObjectVersion>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_objects_snapshot<'life0, 'async_trait>( &'life0 self, object_changes: Vec<TransactionObjectChangesToCommit>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_checkpoints<'life0, 'async_trait>( &'life0 self, checkpoints: Vec<IndexedCheckpoint>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_transactions<'life0, 'async_trait>( &'life0 self, transactions: Vec<IndexedTransaction>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_optimistic_transaction_in_existing_transaction( &self, conn: &mut PgConnection, transaction: OptimisticTransaction, ) -> Result<(), IndexerError>

fn persist_events<'life0, 'async_trait>( &'life0 self, events: Vec<IndexedEvent>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_event_indices<'life0, 'async_trait>( &'life0 self, event_indices: Vec<EventIndex>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_displays<'life0, 'async_trait>( &'life0 self, display_updates: BTreeMap<String, StoredDisplay>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_packages<'life0, 'async_trait>( &'life0 self, packages: Vec<IndexedPackage>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_epoch<'life0, 'async_trait>( &'life0 self, epoch: EpochToCommit, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn advance_epoch<'life0, 'async_trait>( &'life0 self, epoch: EpochToCommit, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_network_total_transactions_by_end_of_epoch<'life0, 'async_trait>( &'life0 self, epoch: u64, ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn refresh_participation_metrics<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn as_any(&self) -> &dyn Any

fn persist_displays_in_existing_transaction( &self, conn: &mut PgConnection, display_updates: Vec<&StoredDisplay>, ) -> Result<(), IndexerError>

fn persist_objects_in_existing_transaction( &self, conn: &mut PgConnection, object_changes: Vec<TransactionObjectChangesToCommit>, ) -> Result<(), IndexerError>

fn update_watermarks_upper_bound<'life0, 'async_trait, E>( &'life0 self, watermark: CommitterWatermark, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where E::Iterator: Iterator<Item: AsRef<str>>, E: 'async_trait + IntoEnumIterator, Self: 'async_trait, 'life0: 'async_trait,

Update the upper bound of the watermarks for the given tables.

fn update_watermarks_lower_bound<'life0, 'async_trait>( &'life0 self, watermarks: Vec<(PrunableTable, u64)>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Updates each watermark entry’s lower bounds per the list of tables and their new epoch lower bounds.

fn get_watermarks<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(Vec<StoredWatermark>, i64), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load all watermark entries from the store, and the latest timestamp from the db.

fn persist_checkpoint_objects<'life0, 'async_trait>( &'life0 self, objects: Vec<CheckpointObjectChanges>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_tx_global_order<'life0, 'async_trait>( &'life0 self, tx_order: Vec<TxGlobalOrder>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn persist_tx_indices<'life0, 'async_trait>( &'life0 self, indices: Vec<TxIndex>, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn prune_table_by_checkpoint_range<'life0, 'life1, 'async_trait>( &'life0 self, table: &'life1 PrunableTable, min_checkpoint: u64, max_checkpoint: u64, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn prune_table_by_tx_range<'life0, 'life1, 'async_trait>( &'life0 self, table: &'life1 PrunableTable, min_tx: u64, max_tx: u64, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn prune_table_by_global_seq_with_limit<'life0, 'life1, 'async_trait>( &'life0 self, table: &'life1 PrunableTable, start: u64, end: u64, limit: i64, ) -> Pin<Box<dyn Future<Output = Result<usize, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Prune table by global_sequence_number range with DELETE LIMIT. Uses DELETE with LIMIT to maintain consistent batch sizes. Returns the number of rows deleted.

fn update_watermark_lowest_unpruned_key<'life0, 'life1, 'async_trait>( &'life0 self, table: &'life1 PrunableTable, lowest_unpruned_key: u64, ) -> Pin<Box<dyn Future<Output = Result<(), IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_watermark_by_entity<'life0, 'async_trait>( &'life0 self, entity: String, ) -> Pin<Box<dyn Future<Output = Result<Option<StoredWatermark>, IndexerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.