iota_indexer::test_utils

Function start_test_indexer_impl

pub async fn start_test_indexer_impl(
    db_url: String,
    reset_db: bool,
    db_init_hook: Option<DBInitHook>,
    rpc_url: String,
    reader_writer_config: IndexerTypeConfig,
    data_ingestion_path: Option<PathBuf>,
    cancel: CancellationToken,
) -> (PgIndexerStore, JoinHandle<Result<(), IndexerError>>)

Starts an indexer reader or writer for testing depending on the reader_writer_config.

Note

For IndexerTypeConfig::Writer when data_ingestion_path is Some, the data ingestion path will be exclusively used to ingest data into the indexer. To force the indexer to sync from the fullnode via gRPC, set data_ingestion_path to None and it will use the rpc_url to stream checkpoints from the fullnode gRPC endpoint.