Module iota_indexer::types

Structs

EventIndex

IndexedCheckpoint

IndexedDeletedObject

IndexedEpochInfo

Represents system state and summary info at the start and end of an epoch. Optional fields are populated at epoch boundary, since they cannot be determined at the start of the epoch.

IndexedEvent

IndexedObject

IndexedPackage

IndexedTransaction

IotaTransactionBlockResponseWithOptions

TxIndex

Enums

DynamicFieldKind

IndexedObjectChange

ObjectStatus

OwnerType

TransactionKind

Functions

owner_to_owner_info

Type Aliases

IndexerResult