iota_indexer::test_utils

Type Alias DBInitHook

pub type DBInitHook = Box<dyn FnOnce(&PgIndexerStore) + Send>;

Type to create hooks to alter initial indexer DB state in tests. Those hooks are meant to be called after DB reset (if it occurs) and before indexer is started.

Example:

let emulate_insertion_order_set_earlier_by_optimistic_indexing: DBInitHook =
    Box::new(move |pg_store: &PgIndexerStore| {
        transactional_blocking_with_retry!(
            &pg_store.blocking_cp(),
            |conn| {
                insert_or_ignore_into!(
                    tx_insertion_order::table,
                    (
                        tx_insertion_order::dsl::tx_digest.eq(digest.inner().to_vec()),
                        tx_insertion_order::dsl::insertion_order.eq(123),
                    ),
                    conn
                );
                Ok::<(), IndexerError>(())
            },
            Duration::from_secs(60)
        )
            .unwrap()
    });

let (_, pg_store, _) = start_simulacrum_grpc_with_write_indexer(
    Arc::new(sim),
    data_ingestion_path,
    None,
    Some("indexer_ingestion_tests_db"),
    Some(emulate_insertion_order_set_earlier_by_optimistic_indexing),
)
.await;

Aliased Type

pub struct DBInitHook(/* private fields */);